+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 
 ### Bug fixes 🐛
+
+- Fix event deduplication [#298](https://github.com/mapbox/mapbox-gl-geocoder/pull/298).
 - Add a paste event handler to ensure that paste events are recognized by the geocoder and trigger searches [#300](https://github.com/mapbox/mapbox-gl-geocoder/pull/300). 
 
 ## v4.4.2